+++ python/examples/dspy-rag-fastapi/backend/app/utils/load.py
@@ -4,8 +4,9 @@ import logging
 import os
 import re
 
-import chromadb
-from chromadb import Documents, EmbeddingFunction, Embeddings
+import openai
+import weaviate
+import weaviate.classes as wvc
 from dotenv import load_dotenv
 
 load_dotenv()
@@ -23,74 +24,58 @@ def chunks(iterable, chunk_size):
         yield iterable[i : i + chunk_size]
 
 
-# Custom Embedding function that supports OPen embeddings
-class OpenAIEmbeddingFunction(EmbeddingFunction[Documents]):
-    """
-    This class is used to get embeddings for a list of texts using OpenAI Python Library.
-    It requires a host url and a model name. The default model name is "text-embedding-ada-002".
-    """
-
-    def __init__(self, model_name: str = "text-embedding-ada-002"):
-        try:
-            import openai
-        except ImportError:
-            raise ValueError("The openai python package is not installed. Please install it")
-
-        self._client = openai.Client()
-        self._model_name = model_name
-
-    def __call__(self, input: Documents) -> Embeddings:
-        """
-        Get the embeddings for a list of texts.
-        Args:
-            input (Documents): A list of texts to get embeddings for.
-        Returns:
-            Embeddings: The embeddings for the texts.
-        Example:
-            >>> openai = OpenAIEmbeddingFunction()
-            >>> texts = ["Hello, world!", "How are you?"]
-            >>> embeddings = openai(texts)
-        """
-
-        embeddings: Embeddings = []
-        # Call OpenAI Embedding API for each document.
-        for document_chunk in chunks(input, chunk_size=10):
-            embedding_response = self._client.embeddings.create(
-                model=self._model_name, input=document_chunk
-            )
-            embeddings.extend([emb.embedding for emb in embedding_response.data])
-
-        return embeddings
-
-
 def load_data() -> None:
     """
-    Loads data from /data/example to Chroma Vector store.
+    Loads data from /data/example to a Weaviate Vector store.
     """
 
     logger.info("Loading data.")
     # Split document into single sentences
-    chunks = []
+    sentence_chunks = []
     with open(f"{DATA_DIR}/example/paul_graham_essay.txt", "r", encoding="utf-8") as file:
         text = file.read()
         sentences = re.split(r"(?<!\w\.\w.)(?<![A-Z][a-z]\.)(?<=\.|\?)\s", text)
         sentences = [sentence.strip() for sentence in sentences if sentence.strip()]
-        chunks.extend(sentences)
+        sentence_chunks.extend(sentences)
 
     logger.info("Creating embeddings.")
-    openai_ef = OpenAIEmbeddingFunction()
-    chunks_embeddings = openai_ef(chunks)
 
-    db = chromadb.PersistentClient(path=f"{DATA_DIR}/chroma_db")
-    chroma_collection = db.get_or_create_collection("quickstart")
-
-    logger.info("Loading data in Chroma.")
-    chroma_collection.add(
-        ids=[f"id{i}" for i in range(1, len(chunks) + 1)],
-        embeddings=chunks_embeddings,
-        documents=chunks,
+    weaviate_client = weaviate.WeaviateClient(
+        embedded_options=weaviate.embedded.EmbeddedOptions(
+            persistence_data_path=f"{DATA_DIR}/weaviate",
+            additional_env_vars={
+                "ENABLE_MODULES": "text2vec-openai",
+                "DEFAULT_VECTORIZER_MODULE": "text2vec-openai",
+            },
+        ),
+        additional_headers={"X-OpenAI-Api-Key": os.getenv("OPENAI_API_KEY")},
     )
-    logger.info("Successfully loaded embeddings in the Chroma.")
+    weaviate_client.connect()
+
+    try:
+        openai_client = openai.Client()
+
+        weaviate_client.collections.delete("paul_graham_essay")
+        collection = weaviate_client.collections.create(
+            "paul_graham_essay", vectorizer_config=wvc.config.Configure.Vectorizer.text2vec_openai()
+        )
+
+        logger.info("Loading data into Weaviate.")
+        with collection.batch.dynamic() as batch:
+            for sentence_chunk in chunks(sentence_chunks, 20):
+                embedding_response = openai_client.embeddings.create(
+                    model="text-embedding-ada-002", input=sentence_chunk
+                )
+                embeddings = [emb.embedding for emb in embedding_response.data]
+                for j, vector in enumerate(embeddings):
+                    batch.add_object(properties={"content": sentence_chunk[j]}, vector=vector)
+        logger.info("Successfully loaded embeddings into Weaviate.")
+
+    except Exception as e:
+        logger.error(f"Failed to load embeddings into Weaviate: {e}")
+
+    finally:
+        weaviate_client.close()
 
 
 if __name__ == "__main__":