+++ js/packages/openinference-instrumentation-langchain/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @arizeai/openinference-instrumentation-langchain
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [921a40c]
+  - @arizeai/openinference-semantic-conventions@0.4.0
+
 ## 0.0.4
 
 ### Patch Changes