+++ js/packages/openinference-instrumentation-llama-index/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @arizeai/openinference-instrumentation-llama-index
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [921a40c]
+  - @arizeai/openinference-semantic-conventions@0.4.0
+
 ## 0.0.2
 
 ### Patch Changes